requirements.txt
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
TA-Lib>=0.4.25
plotly>=5.15.0
requests>=2.31.0
python-dotenv>=1.0.0
alpaca-trade-api>=3.1.1
backtrader>=1.9.78
quantstats>=0.0.62
ipython>=8.0.0